requirements.txt
altair
numpy
pandas
pydeck
streamlit
pinecone-client
pinecone-datasets
datasets
sentence_transformers
torch
pinecone_text